requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==0.19.0
pydantic==2.4.2